Stop word wrapping
Added logic to the addPhraseToDisplay() that places each word of a phrase into a <span></span>.  The span is given a display = "inline-block".  This causes the word to move to a new line if there isn't enough room to display the entire word on the previous line.  Furthermore, the space <li></li> is only added in between each span.